✨: json-server를 기반으로 중복검사 구현
