+++ mocking/db/db.json
@@ -1,28 +1,9 @@
 {
   "users": [
     {
-      "id": 1,
       "username": "testUser",
       "email": "testUser@example.com",
       "password": "1234"
-    },
-    {
-      "id": 2,
-      "username": "testUser",
-      "email": "testUser@example.com",
-      "password": "1234"
-    },
-    {
-      "username": "testUser",
-      "email": "testUser@example.com",
-      "password": "1234",
-      "id": 3
-    },
-    {
-      "username": "testUser",
-      "email": "testUser@example.com",
-      "password": "1234",
-      "id": 3
     }
   ],
   "names": [
