+++ mocking/json.mocking.server.js
@@ -1,10 +1,9 @@
 import jsonServer from 'json-server';
 import path from 'path';
 import { fileURLToPath } from 'url';
-// import createRoutes from './Routes/createRoutes.js';
-// import signupHandler from './function/signup.handler.js';
-// import getNameHandler from './function/getName.handler.test.js';
-// import getUserHandler from './function/getUser.handler.test copy.js';
+import createRoutes from './Routes/createRoutes.js';
+import { postSignUpHandler } from './function/PostHandler.js';
+import { getTestNameHandler } from './function/GetHandler.js';
 // ESM에서 __dirname을 설정하는 방법
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -12,40 +11,22 @@ const __dirname = path.dirname(__filename);
 const server = jsonServer.create();
 
 // JSON 파일 경로 설정
-const middlewares = jsonServer.defaults(); //json-server를 기준으로 설정
-const routes = jsonServer.router(path.join(__dirname, 'db', 'db.json'));
-// const test = jsonServer.router(path.join(__dirname, 'db', 'db.user.json'));
-const rewriter = jsonServer.rewriter(path.join(__dirname, 'routes.json')); // routes.json을 로드
+
+const router = jsonServer.router(path.join(__dirname, 'db', 'db.json')); // db.json 파일 경로
+const middlewares = jsonServer.defaults();
+const rewriter = jsonServer.rewriter(path.join(__dirname, 'routes.json'));
 
 server.use(middlewares);
 server.use(jsonServer.bodyParser);
-server.use(rewriter); // 라우트 재작성 미들웨어 추가
-server.use(routes);
+server.use(rewriter);
 
-// 커스텀 라우트 예제
-// createRoutes(server, 'get', '/api/user/signup', getUserHandler(test));
-// createRoutes(server, 'get', '/api/test/names', getNameHandler(routes));
+// 커스텀 라우트
 
-/** Custom Routing */
-//json-server를 기준으로 가져오는 코드임 즉, json-server를 읽는 코드
-server.get('/api', (req, res) => {
-  res.send(routes.get('api').value());
-});
+createRoutes(server, 'post', '/api/user/signup', postSignUpHandler(router));
+createRoutes(server, 'get', '/api/test/name', getTestNameHandler(router));
 
-// POST 요청을 처리하는 커스텀 라우트
-server.post('/api/user/signup', (req, res) => {
-  const users = routes.db.get('users');
-  const user = req.body;
-
-  // 중복 사용자 체크
-  const existingUser = users.find({ username: user.username }).value(); // user.username으로 수정
-  if (existingUser) {
-    return res.status(400).json({ error: 'User already exists' });
-  }
-
-  users.push(user).write(); // 사용자 추가
-  res.status(201).json(user);
-});
+// 기존 json-server 라우터 사용
+server.use(router);
 
 const PORT = 3001;
 server.listen(PORT, () => {
