+++ mocking/test/db/routes.json
@@ -1,0 +1,4 @@
+{
+  "/api/user/signup": "/users",
+  "/api/test/name": "/names"
+}
